... but need to move tests for #71 back under failing for 3.x due to diff in yaml engine
